Change price from uint256 to uint96 (#110)

+++ tests/mocks/MockMinter.sol
@@ -23,7 +23,7 @@ contract MockMinter is BaseMinter {
         _mint(edition, mintId, quantity, affiliate);
     }
 
-    function setPrice(uint256 price_) external {
+    function setPrice(uint96 price_) external {
         _currentPrice = price_;
     }
